+++ io-core/src/jvmMain/kotlin/io/github/fukusaka/keel/io/NativeBuf.kt
@@ -44,6 +44,14 @@ actual class NativeBuf internal actual constructor(actual val capacity: Int) {
         writerIndex += length
     }
 
+    actual fun writeAsciiString(src: String, srcOffset: Int, length: Int) {
+        require(length <= writableBytes) { "length $length exceeds writableBytes $writableBytes" }
+        for (i in 0 until length) {
+            buf.put(writerIndex + i, src[srcOffset + i].code.toByte())
+        }
+        writerIndex += length
+    }
+
     actual fun readByte(): Byte = buf.get(readerIndex++)
 
     actual fun getByte(index: Int): Byte = buf.get(index)